refactor: remove model.rs and refactor prompt
