--- llama-cli/src/commands/model.rs
@@ -1,5 +1,0 @@
-    pub struct Cmd {
-        /// Where to load the model path from
-        #[arg(long, short = 'm')]
-        model_path: String,
-    },
